% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waapihk_style.R
\name{waapihk_style}
\alias{waapihk_style}
\title{Add waapihk theme to ggplot chart}
\usage{
waapihk_style()
}
\description{
This function allows you to add the waapihk theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
waapihk_style()
}
\keyword{waapihk_style}
